# WAL: Custom Encoding for `Envelope`