… Cargo.lock
… crates/partition-store/src/protobuf_types.rs
… crates/wal-protocol/src/lib.rs