WAL: Custom Encoding for `Envelope`

This PR enables `Envelope` to apply custom encoding strategies for individual Command variants, selecting the most efficient encoding available for each case.

The design is forward-compatible, allowing future changes to the encoding of specific variants without breaking backward compatibility.

## Benches
Benchmarks shows a big improvement for variants that uses protobuf (e.g invoke-cmd in this example).
While also unfortunately shows a small regression in performance for variants that still uses flexbuffers internally.

Those variants will eventually switch to bilrost overtime, so I think even if we have a slight decrease in performance right now
this changeset will allow us to improve this later on.

*NOTE:* To run the benchmarks locally make sure to modify the envelope `StorageEncode` implementation
to actually use `Custom` encoding

## Before
<details>

<summary>Benchmarks results before switching to custom encoding</summary>

```
     Running benches/replicated_loglet_serde.rs (target/release/deps/replicated_loglet_serde-5dda2eb5d1331e02)
Gnuplot not found, using plotters backend
replicated-loglet-serde.invoke-cmd/serialize-append
                        time:   [58.196 µs 58.248 µs 58.305 µs]
                        change: [+64.298% +64.636% +64.929%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 50 measurements (12.00%)
  1 (2.00%) low severe
  1 (2.00%) low mild
  4 (8.00%) high mild
replicated-loglet-serde.invoke-cmd/deserialize-append
                        time:   [1.9243 µs 1.9267 µs 1.9295 µs]
                        change: [-4.3010% -4.1884% -4.0626%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 50 measurements (8.00%)
  2 (4.00%) high mild
  2 (4.00%) high severe
replicated-loglet-serde.invoke-cmd/serialize-store
                        time:   [58.438 µs 58.475 µs 58.517 µs]
                        change: [+62.380% +62.647% +62.876%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 50 measurements (4.00%)
  1 (2.00%) low mild
  1 (2.00%) high mild

replicated-loglet-serde.invoker-effect/serialize-append
                        time:   [36.209 µs 36.256 µs 36.319 µs]
                        change: [-8.3523% -8.1153% -7.9057%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 50 measurements (8.00%)
  1 (2.00%) high mild
  3 (6.00%) high severe
replicated-loglet-serde.invoker-effect/deserialize-append
                        time:   [1.9168 µs 1.9208 µs 1.9262 µs]
                        change: [+1.3047% +1.6301% +1.8847%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 50 measurements (6.00%)
  1 (2.00%) high mild
  2 (4.00%) high severe
replicated-loglet-serde.invoker-effect/serialize-store
                        time:   [36.556 µs 36.603 µs 36.663 µs]
                        change: [-8.4178% -8.2808% -8.1429%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 50 measurements (10.00%)
  2 (4.00%) high mild
  3 (6.00%) high severe
```

</details>

### After
<details>

<summary>Benchmarks results after switching to custom encoding</summary>

```
     Running benches/replicated_loglet_serde.rs (target/release/deps/replicated_loglet_serde-fdc6dca006aa05a2)
Gnuplot not found, using plotters backend
replicated-loglet-serde.invoke-cmd/serialize-append
                        time:   [35.439 µs 35.468 µs 35.500 µs]
                        change: [-39.114% -39.039% -38.965%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 50 measurements (6.00%)
  2 (4.00%) high mild
  1 (2.00%) high severe
replicated-loglet-serde.invoke-cmd/deserialize-append
                        time:   [1.8101 µs 1.8129 µs 1.8161 µs]
                        change: [-6.0752% -5.9156% -5.7532%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 50 measurements (16.00%)
  1 (2.00%) low mild
  3 (6.00%) high mild
  4 (8.00%) high severe
replicated-loglet-serde.invoke-cmd/serialize-store
                        time:   [35.910 µs 35.944 µs 35.980 µs]
                        change: [-38.591% -38.463% -38.312%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 50 measurements (6.00%)
  2 (4.00%) high mild
  1 (2.00%) high severe

replicated-loglet-serde.invoker-effect/serialize-append
                        time:   [39.387 µs 39.453 µs 39.520 µs]
                        change: [+8.3225% +8.5442% +8.7673%] (p = 0.00 < 0.05)
                        Performance has regressed.
replicated-loglet-serde.invoker-effect/deserialize-append
                        time:   [1.8490 µs 1.8517 µs 1.8547 µs]
                        change: [-3.8572% -3.6454% -3.4561%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 50 measurements (8.00%)
  4 (8.00%) high mild
replicated-loglet-serde.invoker-effect/serialize-store
                        time:   [39.832 µs 39.868 µs 39.905 µs]
                        change: [+8.6906% +8.8668% +9.0299%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 50 measurements (10.00%)
  2 (4.00%) low mild
  3 (6.00%) high severe
```
</details>