@muhamadazmy 2025-05-12
@AhmedSoliman I still need to run some benches to make sure this does not affect performance (in flexbuffers variants cases).